Fixed Windows build
- missing include in spectrum/types.cpp
- missing include in zx82snapshotreader.cpp
- missing include in pagingmemory.h
- misnamed function cal in notification.cpp
- implicit cast in Cpu::memorySize() [VS warning]

+++ src/cpu.h
@@ -65,7 +65,7 @@ public:
 
     [[nodiscard]] inline int memorySize() const
     {
-        return m_memory->addressableSize();
+        return static_cast<int>(m_memory->addressableSize());
     }
 
     inline void setClockSpeed(unsigned long long hz)
